api: finish refactor
