+++ conftest.py
@@ -3,23 +3,25 @@ import os
 import pytest
 import socketio
 
-os.environ['DATABASE_NAME'] = 'test'
+from hearts import create_app
+from hearts import mongo
+
+os.environ['DATABASE_URL'] = 'mongodb://127.0.0.1:27017/test'
+app = create_app()
 
 
 @pytest.fixture(scope="session")
 def api(request):
-    from app import app
     app.config['TESTING'] = True
     return app.test_client()
 
 
 @pytest.fixture(scope="session")
 def socket(request):
-    from app import app
     return socketio.test_client(app)
 
 
 @pytest.fixture(scope="module")
-def db(request):
-    from app import db_client
-    return db_client
+def db(api, request):
+    with app.app_context():
+        return mongo.db
