+++ env.template
@@ -1,3 +1,3 @@
 HOST=0.0.0.0
 
-DATABASE_URL=mongodb://localhost:27017/
+DATABASE_URL=mongodb://localhost:27017/hearts
