+++ hearts/api/strings.py
