+++ hearts/static/jquery.js
