+++ requirements.txt
@@ -1,3 +1,4 @@
+Flask-PyMongo==0.4.1
 Flask-SocketIO==2.7.1
 Flask==0.11.1
 Jinja2==2.8
