+++ tox.ini
@@ -1,6 +1,6 @@
 [flake8]
 max-line-length = 120
-exclude = venv/*
+exclude = venv/*,hearts/api/__init__.py
 
 [pytest]
 addopts = -v --cov-report term-missing --cov .
